refactoring bundle command changes as per review comments
