+++ core/src/main/kotlin/run/qontract/core/git/GitCommand.kt
@@ -21,5 +21,4 @@ interface GitCommand {
     fun relativeGitPath(newerContractPath: String): Pair<GitCommand, String>
     fun fileIsInGitDir(newerContractPath: String): Boolean
     fun inGitRootOf(contractPath: String): GitCommand
-    fun repoName(): String
 }
